- remove bug in i2s driver resulting in overwriting DMA descriptor while initialization leading to wrong initial sync - change some function return values from int8_t to int32_t - improve syncing   o try to mimic original algorithm used by badaix - latency_buffer_full() call in player sporadically returns negative value   which led to a unnecessary vTaskDelay --> resolved - use esp_timer_get_time() instead of gettimeofday() for timestamping - enable IRAM optimazations for WIFI - decrease WIFI RX buffers

+++ components/libmedian/include/MedianFilter.h
@@ -36,7 +36,7 @@ typedef struct {
 
 int MEDIANFILTER_Init(sMedianFilter_t *medianFilter);
 int64_t MEDIANFILTER_Insert(sMedianFilter_t *medianFilter, int64_t sample);
-uint8_t MEDIANFILTER_isFull(sMedianFilter_t *medianFilter);
+uint32_t MEDIANFILTER_isFull(sMedianFilter_t *medianFilter);
 
 #ifdef __cplusplus
 }
